+++ pkg/ui/bucket.go
@@ -6,17 +6,18 @@ package ui
 import (
 	"html/template"
 	"net/http"
+	"path"
 	"time"
 
 	"github.com/go-kit/kit/log"
 	"github.com/prometheus/common/route"
+	"github.com/thanos-io/thanos/pkg/component"
 	extpromhttp "github.com/thanos-io/thanos/pkg/extprom/http"
 )
 
 // Bucket is a web UI representing state of buckets as a timeline.
 type Bucket struct {
 	*BaseUI
-	flagsMap map[string]string
 	// Unique Prometheus label that identifies each shard, used as the title. If
 	// not present, all labels are displayed externally as a legend.
 	Label       string
@@ -27,10 +28,9 @@ type Bucket struct {
 
 func NewBucketUI(logger log.Logger, label string, flagsMap map[string]string) *Bucket {
 	return &Bucket{
-		BaseUI:   NewBaseUI(logger, "bucket_menu.html", queryTmplFuncs()),
-		Blocks:   "[]",
-		Label:    label,
-		flagsMap: flagsMap,
+		BaseUI: NewBaseUI(logger, "bucket_menu.html", queryTmplFuncs(), flagsMap, component.Bucket),
+		Blocks: "[]",
+		Label:  label,
 	}
 }
 
@@ -42,6 +42,15 @@ func (b *Bucket) Register(r *route.Router, ins extpromhttp.InstrumentationMiddle
 
 	r.Get("/", instrf("root", b.root))
 	r.Get("/static/*filepath", instrf("static", b.serveStaticAsset))
+	// Make sure that "<path-prefix>/new" is redirected to "<path-prefix>/new/" and
+	// not just the naked "/new/", which would be the default behavior of the router
+	// with the "RedirectTrailingSlash" option (https://godoc.org/github.com/julienschmidt/httprouter#Router.RedirectTrailingSlash),
+	// and which breaks users with a --web.route-prefix that deviates from the path derived
+	// from the external URL.
+	r.Get("/new", func(w http.ResponseWriter, r *http.Request) {
+		http.Redirect(w, r, path.Join(GetWebPrefix(b.logger, b.flagsMap, r), "new")+"/", http.StatusFound)
+	})
+	r.Get("/new/*filepath", instrf("react-static", b.serveReactUI))
 }
 
 // Handle / of bucket UIs.