+++ pkg/ui/query.go
@@ -26,8 +26,6 @@ type Query struct {
 	*BaseUI
 	storeSet *query.StoreSet
 
-	flagsMap map[string]string
-
 	cwd   string
 	birth time.Time
 	reg   prometheus.Registerer
@@ -49,9 +47,8 @@ func NewQueryUI(logger log.Logger, reg prometheus.Registerer, storeSet *query.St
 		cwd = "<error retrieving current working directory>"
 	}
 	return &Query{
-		BaseUI:   NewBaseUI(logger, "query_menu.html", queryTmplFuncs()),
+		BaseUI:   NewBaseUI(logger, "query_menu.html", queryTmplFuncs(), flagsMap, component.Query),
 		storeSet: storeSet,
-		flagsMap: flagsMap,
 		cwd:      cwd,
 		birth:    time.Now(),
 		reg:      reg,
@@ -83,6 +80,16 @@ func (q *Query) Register(r *route.Router, ins extpromhttp.InstrumentationMiddlew
 	r.Get("/status", instrf("status", q.status))
 
 	r.Get("/static/*filepath", instrf("static", q.serveStaticAsset))
+	// Make sure that "<path-prefix>/new" is redirected to "<path-prefix>/new/" and
+	// not just the naked "/new/", which would be the default behavior of the router
+	// with the "RedirectTrailingSlash" option (https://godoc.org/github.com/julienschmidt/httprouter#Router.RedirectTrailingSlash),
+	// and which breaks users with a --web.route-prefix that deviates from the path derived
+	// from the external URL.
+	r.Get("/new", func(w http.ResponseWriter, r *http.Request) {
+		http.Redirect(w, r, path.Join(GetWebPrefix(q.logger, q.flagsMap, r), "new")+"/", http.StatusFound)
+	})
+	r.Get("/new/*filepath", instrf("react-static", q.serveReactUI))
+
 	// TODO(bplotka): Consider adding more Thanos related data e.g:
 	// - What store nodes we see currently.
 	// - What sidecars we see currently.