+++ pkg/ui/rule.go
@@ -16,14 +16,13 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/common/route"
 	"github.com/prometheus/prometheus/rules"
+	"github.com/thanos-io/thanos/pkg/component"
 	extpromhttp "github.com/thanos-io/thanos/pkg/extprom/http"
 	thanosrule "github.com/thanos-io/thanos/pkg/rule"
 )
 
 type Rule struct {
 	*BaseUI
-
-	flagsMap map[string]string
 
 	ruleManager *thanosrule.Manager
 	queryURL    string
@@ -32,8 +31,7 @@ type Rule struct {
 
 func NewRuleUI(logger log.Logger, reg prometheus.Registerer, ruleManager *thanosrule.Manager, queryURL string, flagsMap map[string]string) *Rule {
 	return &Rule{
-		BaseUI:      NewBaseUI(logger, "rule_menu.html", ruleTmplFuncs(queryURL)),
-		flagsMap:    flagsMap,
+		BaseUI:      NewBaseUI(logger, "rule_menu.html", ruleTmplFuncs(queryURL), flagsMap, component.Rule),
 		ruleManager: ruleManager,
 		queryURL:    queryURL,
 		reg:         reg,
@@ -163,6 +161,15 @@ func (ru *Rule) Register(r *route.Router, ins extpromhttp.InstrumentationMiddlew
 	r.Get("/rules", instrf("rules", ru.rules))
 
 	r.Get("/static/*filepath", instrf("static", ru.serveStaticAsset))
+	// Make sure that "<path-prefix>/new" is redirected to "<path-prefix>/new/" and
+	// not just the naked "/new/", which would be the default behavior of the router
+	// with the "RedirectTrailingSlash" option (https://godoc.org/github.com/julienschmidt/httprouter#Router.RedirectTrailingSlash),
+	// and which breaks users with a --web.route-prefix that deviates from the path derived
+	// from the external URL.
+	r.Get("/new", func(w http.ResponseWriter, r *http.Request) {
+		http.Redirect(w, r, path.Join(GetWebPrefix(ru.logger, ru.flagsMap, r), "new")+"/", http.StatusFound)
+	})
+	r.Get("/new/*filepath", instrf("react-static", ru.serveReactUI))
 }
 
 // AlertStatus bundles alerting rules and the mapping of alert states to row classes.